backend/requirements.txt
pip>=25.3
setuptools>=75.0.0
wheel

fastapi==0.95.0
uvicorn[standard]==0.21.0
python-multipart==0.0.5
python-dotenv==1.0.0
httpx==0.24.0
pydantic==1.10.24
openai==1.3.0
psycopg[binary]==3.2.3
pgvector==0.3.1
supabase==2.4.0
pillow>=10.3.0,<11.0.0
pytesseract==0.3.10